+++ NeonEngine/NeonEngine/App/Input/Mouse.h
@@ -24,7 +24,7 @@ namespace Neon {
 			~Mouse();
 
 			inline void SetButton(const int button, const bool state) { m_buttons[button] = state; }
-			inline void SetPosition(int x, int y) { m_pos.x = x; m_pos.y = y; }
+			inline void SetPosition(double x, double y) { m_pos.x = (float)x; m_pos.y = (float)y; }
 
 			inline const bool GetButton(const int button) const { return m_buttons[button]; }
 			inline const glm::vec2 &GetPosition() const { return m_pos; }
